% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statgl_colors.R
\name{scale_color_statgl}
\alias{scale_color_statgl}
\title{Statistics Greenland color scale}
\usage{
scale_color_statgl(palette = "main", discrete = TRUE,
  reverse = FALSE, ...)
}
\arguments{
\item{palette}{Character string with name of palette. Can be changed to \code{"dawn"}, etc. Defaults to \code{"main"}}

\item{discrete}{Generate discrete palette if \code{TRUE}}

\item{reverse}{Generate reverse palette if \code{TRUE}}

\item{...}{Parameters to \code{discrete_scale} or \code{scale_color_gradientn}}
}
\description{
Generates a color scale compatible with ggplot2. Uses colors defined by the GS design manual.
}
\examples{
gg_example <- ggplot2::qplot(displ, hwy, data = ggplot2::mpg, color = class)

gg_example + scale_color_statgl()
gg_example + scale_color_statgl(palette = "aurora", reverse = TRUE)
}
