% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statgl_colors.R
\name{scale_fill_statgl}
\alias{scale_fill_statgl}
\title{Statistics Greenland fill scale}
\usage{
scale_fill_statgl(palette = "main", discrete = TRUE, reverse = FALSE,
  ...)
}
\arguments{
\item{palette}{Character string with name of palette. Can be changed to \code{"dawn"}, etc. Defaults to \code{"main"}}

\item{discrete}{Generate discrete palette if \code{TRUE}}

\item{reverse}{Generate reverse palette if \code{TRUE}}

\item{...}{Parameters to \code{discrete_scale} or \code{scale_color_gradientn}}
}
\description{
Generates a fill scale compatible with ggplot2. Uses colors defined by the GS design manual.
}
\examples{
gg_example <- ggplot2::qplot(manufacturer, fill = manufacturer, data = ggplot2::mpg)

gg_example + scale_fill_statgl()
gg_example + scale_fill_statgl(palette = "aurora")
}
