% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statgl_report.R
\name{statgl_report}
\alias{statgl_report}
\title{Render statgl report}
\usage{
statgl_report(output = "html", ...)
}
\arguments{
\item{output}{Format of report. One of \code{"html"}, \code{"docx"}, or
\code{"pretty_html"}.}

\item{...}{Additional arguments passed to the underlying rmarkdown
output format (e.g. \code{\link[rmarkdown:html_document]{rmarkdown::html_document()}}).}
}
\value{
An R Markdown output format object suitable for the \code{output}
field of a YAML header.
}
\description{
Wrapper function for rmarkdown::render(), to render html pages with statgl
design.
}
\details{
This is a development version, and design of the output will change!
}
\examples{
statgl_report()
}
