% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statgl_search.R
\name{statgl_search}
\alias{statgl_search}
\title{Search Statistics Greenland API}
\usage{
statgl_search(
  query = "",
  path = "",
  lang = "en",
  api_url = statgl_api_url(),
  returnclass = "tibble"
)
}
\arguments{
\item{query}{Search query}

\item{path}{Subfolder of API to search, e.g. \code{"BE"} for population}

\item{lang}{Language. Defaults to \code{"en"}, and must be one of
\code{c("en", "kl", "da")} for Statistics Greenland PX API}

\item{api_url}{Base API URL.}

\item{returnclass}{Return class from search query. Must be one of
\code{"tibble"} or \code{"list"}. Defaults to tibble. Returns to default list
from API when list.}
}
\value{
When \code{returnclass = "tibble"} (default), a tibble of matching
tables and folders. When \code{returnclass = "list"}, the raw parsed API
response.
}
\description{
Do a search of Statistics Greenland API. Will work for other PXWEB APIs, when
custom URL is provided in \code{api_url}
}
\examples{
\donttest{
statgl_search("population")
statgl_search("inuit", lang = "kl")
}
}
