% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_statgl.R
\name{theme_statgl}
\alias{theme_statgl}
\title{Statistics Greenland ggplot2 theme}
\usage{
theme_statgl(base_size = 12, base_family = NULL, ...)
}
\arguments{
\item{base_size}{Base font size}

\item{base_family}{Base font family}

\item{...}{Parameters to \code{ggplot2::theme}}
}
\description{
A ggplot2 theme following design manual of Statistics Greenland.
}
\examples{
ggplot2::qplot(mpg, data = mtcars) + theme_statgl()
}
