% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statgl_fetch.R
\name{px_top}
\alias{px_top}
\alias{px_all}
\alias{px_agg}
\title{Fetch helpers}
\usage{
px_top(top_n = 1)

px_all(pattern = "*")

px_agg(agg_file, ...)
}
\arguments{
\item{top_n}{Numeric. Requests the first n values of a variable, or the last
n values if variable is time.}

\item{pattern}{Character. Requests variables matching a wildcard string. E.g.
\code{"\*0"} returns all variables ending in 0, \code{"*"} returns all
values}

\item{agg_file}{Aggregation file used to aggregate values in pxweb API}

\item{...}{Values selected from aggregation}
}
\description{
Helper functions to make custom filters in pxweb API queries:
}
\details{
* [px_top()]: API returns the the top n values provided
* [px_all()]: API returns all values, matching a wildcard pattern
* [px_agg()]; Instructs API to use aggregations, and which values to select.
}
