% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statgl_fetch.R
\name{statgl_fetch}
\alias{statgl_fetch}
\title{Retrieve statbank data via URL}
\usage{
statgl_fetch(
  x,
  ...,
  .col_code = FALSE,
  .val_code = FALSE,
  .eliminate_rest = TRUE,
  url = NULL
)
}
\arguments{
\item{x}{API url of statbank matrix}

\item{...}{Selection queries for variables}

\item{.col_code}{\code{TRUE}/\code{FALSE}. Display column names as code.}

\item{.val_code}{\code{TRUE}/\code{FALSE}. Display cell values as code.}

\item{.eliminate_rest}{\code{TRUE}/\code{FALSE}. If \code{FALSE}, retrive all selections for remaining variables in table (experimental).}

\item{url}{deprecated}
}
\value{
return a dataframe / tibble with the data from the registry
}
\description{
This funtion retrive data from statistics banks via URL and hidden quiestions in a qury

The variables is displayed by their text_code by default. This can be changed
by setting .col_code = TRUE.

The selection of variables should be given using their code values and not text values.

These can be found by using the statgl_meta() function prior to using statgl_fetch()
}
\details{
The variables is displayed by their text_code by default. This can be changed
by setting .col_code = TRUE.

The selection of variables should be given using their code values and not text values.

These can be found by using the statgl_meta() function prior to using statgl_fetch()
}
\examples{
statgl_fetch("BEXSTA")
statgl_fetch("BEXSTA")
statgl_fetch("BEXSTA", gender = c("M", "K"), time = 2010:2020)
statgl_fetch(statgl_url("BEXSTA"), time = px_top(1), age = px_all("*0"))
}
