% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statgl_fetch.R
\name{statgl_fetch}
\alias{statgl_fetch}
\title{Retrieve statbank data via URL}
\usage{
statgl_fetch(
  url,
  ...,
  .col_code = FALSE,
  .val_code = FALSE,
  .eliminate_rest = TRUE
)
}
\arguments{
\item{url}{API url of statbank matrix}

\item{...}{Selection queries for variables}

\item{.col_code}{\code{TRUE}/\code{FALSE}. Display column names as code.}

\item{.val_code}{\code{TRUE}/\code{FALSE}. Display cell values as code.}

\item{.eliminate_rest}{\code{TRUE}/\code{FALSE}. If \code{FALSE}, retrive all selections for remaining variables in table (experimental).}
}
\description{
Retrieve statbank data via URL
}
\examples{
statgl_fetch("BEXSTA")
statgl_fetch(statgl_url("BEXSTA"))
statgl_fetch(statgl_url("BEXSTA"), gender = c("M", "K"), time = 2010:2020)
statgl_fetch(statgl_url("BEXSTA"), time = px_top(1), age = px_all("*0"))
}
