% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statgl_meta.R
\name{statgl_meta}
\alias{statgl_meta}
\title{Retrieve metadata from Statbank}
\usage{
statgl_meta(url, returnclass = "tibble")
}
\arguments{
\item{url}{API url of Statbank table}

\item{returnclass}{Returns list with tibble if \code{"tibble"} (default).}
}
\description{
Retrieves metadata from pxweb API url.
}
\examples{
statgl_meta(statgl_url("BEXSTA"))
}
