% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statgl_report.R
\name{statgl_report}
\alias{statgl_report}
\title{Render statgl report}
\usage{
statgl_report(output = "html", ...)
}
\arguments{
\item{output}{Format of report. Must be \code{"html"} in current release.}

\item{...}{Additional arguments passed to rmarkdown::html_document}
}
\description{
Wrapper function for rmarkdown::render(), to render html pages with statgl
design.
}
\details{
This is a development version, and design of the output will change!
}
\examples{
statgl_report()
}
