% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statgl_search.R
\name{statgl_search}
\alias{statgl_search}
\title{Search Statistics Greenland API}
\usage{
statgl_search(query, lang = "en", path = "",
  api_url = paste0("https://bank.stat.gl:443/api/v1/", lang,
  "/Greenland/", path))
}
\arguments{
\item{query}{Search query}

\item{lang}{Language. Defaults to \code{"en"}, and must be one of
\code{c("en", "kl", "da")} for Statistics Greenland PX API}

\item{path}{Subfolder of API to search, e.g. \code{"BE"} for population}

\item{api_url}{Base API URL.}
}
\description{
Do a search of Statistics Greenland API. Will work for other PXWEB APIs, when
custom URL is provided in \code{api_url}
}
\examples{
statgl_search("population")
statgl_search("inuit", lang = "kl")
}
