% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statgl_table.R
\name{statgl_table}
\alias{statgl_table}
\title{Automatic statgl table}
\usage{
statgl_table(df, year_col, ...)
}
\arguments{
\item{df}{A data frame}

\item{year_col}{Name of year column. Disregards formatiing at this num column.}

\item{...}{Parameters to \code{format}}
}
\description{
Concvenience function to quickly write html tables. Returns a kable with automatic alignments and formats.
}
\examples{
statgl_table(ggplot2::mpg, year_col = year)
}
