% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statgl_url.R
\name{statgl_url}
\alias{statgl_url}
\title{Retrieve URL of px table}
\usage{
statgl_url(
  table_id,
  lang = get_language(substr(table_id, 3, 3)),
  api_url = paste0("https://bank.stat.gl:443/api/v1/", lang, "/Greenland/")
)
}
\arguments{
\item{table_id}{ID of table}

\item{lang}{API language of query. Defaults to \code{"en"}}

\item{api_url}{URL of Statbank API. Defaults to Statistics Greenland API URL}
}
\description{
Searches the statbank for a table of a given ID, and returns the URL if
found. Only works on PX tables from Statistics Greenland.
}
\examples{
statgl_url("BEXST1")
statgl_url("BEXST1", lang = "kl")
}
