% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/word_to_qmd_addin.R
\name{word_to_qmd}
\alias{word_to_qmd}
\title{Convert a Word document to Quarto (QMD) (non-interactive)}
\usage{
word_to_qmd(in_file)
}
\arguments{
\item{in_file}{Path to .docx}
}
\description{
Convert a Word document to Quarto (QMD) (non-interactive)
}
